feat: add support for run public api checks and add run module
