+++ package.json
@@ -69,7 +69,6 @@
     "axios": "0.21.2",
     "builtin-modules": "3.2.0",
     "chalk": "4.1.1",
-    "cli-fuzzy-search": "2.0.0",
     "cli-spinners": "2.6.1",
     "conf": "8.0.0",
     "consola": "2.15.0",
