+++ src/modules/run/index.js
@@ -1,0 +1,7 @@
+const apiCheck = require('./api-check')
+const browserCheck = require('./browser-check')
+
+module.exports = {
+  apiCheck,
+  browserCheck
+}
